小整理,使用体验细节

+++ CUI/GUI/Label.cpp
@@ -48,7 +48,7 @@ void Label::Update()
 	auto d2d = this->ParentForm->Render;
 	auto size = this->ActualSize();
 	auto font = this->Font;
-	float clipW = last_width > size.cx ? (float)last_width : FLT_MAX;
+	float clipW = last_width > static_cast<float>(size.cx) ? last_width : FLT_MAX;
 	this->BeginRender(clipW, FLT_MAX);
 	{
 		if (this->Image)
@@ -59,9 +59,9 @@ void Label::Update()
 	}
 	if (!this->Enable)
 	{
-		float w = last_width > size.cx ? (float)last_width : (float)size.cx;
-		d2d->FillRect(0, 0, w, size.cy, { 1.0f ,1.0f ,1.0f ,0.5f });
+		float w = last_width > static_cast<float>(size.cx) ? last_width : static_cast<float>(size.cx);
+		d2d->FillRect(0, 0, w, static_cast<float>(size.cy), { 1.0f ,1.0f ,1.0f ,0.5f });
 	}
 	this->EndRender();
-	last_width = size.cx;
+	last_width = static_cast<float>(size.cx);
 }
