+++ CUI/GUI/LinkLabel.cpp
@@ -18,7 +18,7 @@ void LinkLabel::Update()
 	auto size = this->ActualSize();
 	auto font = this->Font;
 
-	float clipW = last_width > size.cx ? (float)last_width : FLT_MAX;
+	float clipW = last_width > static_cast<float>(size.cx) ? last_width : FLT_MAX;
 	this->BeginRender(clipW, FLT_MAX);
 	{
 		if (this->Image)
@@ -40,12 +40,12 @@ void LinkLabel::Update()
 
 	if (!this->Enable)
 	{
-		float w = last_width > size.cx ? (float)last_width : (float)size.cx;
-		d2d->FillRect(0, 0, w, size.cy, { 1.0f ,1.0f ,1.0f ,0.5f });
+		float w = last_width > static_cast<float>(size.cx) ? last_width : static_cast<float>(size.cx);
+		d2d->FillRect(0, 0, w, static_cast<float>(size.cy), { 1.0f ,1.0f ,1.0f ,0.5f });
 	}
 
 	this->EndRender();
-	last_width = size.cx;
+	last_width = static_cast<float>(size.cx);
 }
 
 CursorKind LinkLabel::QueryCursor(int xof, int yof)
